Move .envrc to .envrc.template
`.envrc` is usually a user’s own thing. For me for instance it doesn’t
make sense to use this one because I use NixOS and I don’t need Stack.
Overridding `ghci` to `stack ghci` is not what I’m looking for.
But on the other hand I put `use nix` to `.envrc` which makes direnv to
automatically enter the project’s nix-shell configuration I just added
for myself. So I have to change `.envrc` but should pay attention to
not commit it accidentally since it’s controlled by Git.

`.envrc.template` would allow a developer to just copy-paste it and
modify if necessary. Or to just create a symlink to this template if
it’s fine as-is.

But whatever one puts into `.envrc` is their personal business.
Git would just ignore it.